refactor: 타입 파일로 분리 (#29)
